requirements.txt
argparse
google-generativeai
